[FIX] analytic: save analytic_distribution in wizard views
Steps to reproduce the issue:

1. Make sure you have access to the Analytic Distribution fields
2. Create an expense
3. Open the Split Expense Wizard and add Analytic Distribution to the new Expenses
    - Click out of the Analytic Distribution widget to close it, don't use the close icon
4. Finish the splitting process
5. On the Tree view that follows, the Analytic Distribution fields are empty

Explanation:

`onWindowClick` checks if the `analytic_distirbution` widget should close. If the conditions are met, `forceCloseEditor` is called and the data is saved.
In this case, one of the `selectors` is retrieved by `ev.target.closest`, namely `.modal:not(.o_inactive_modal)`, therefore not meeting the conditions to call `forceCloseEditor`, but the widget still closes.

Fix reasoning:

We want `forceCloseEditor` to be called in this situation, since the user is not clicking on an element related to `analytic_distribution`. We can use  `o_act_window` to filter Search More modals from Wizards, as it is absent from the former.

opw-4001757

closes odoo/odoo#178123

X-original-commit: 9ac874046ab4f9c6df58e4d039959015efa55df4
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Stroobant Paul (stpa) <stpa@odoo.com>